backend/requirements.txt
motor
dotenv
pillow
numpy
fastapi
deepface
pydantic
multipart
uvicorn
ultralytics
redis